bugfix: don't cast to ints

+++ pcl.pyx
@@ -117,7 +117,7 @@ cdef class PointCloud:
 
     def to_array(self):
         #FIXME: this could be done more efficinetly, i'm sure
-        return np.array(self.to_list(), dtype=np.int32)
+        return np.array(self.to_list(), dtype=np.float32)
 
     def from_list(self, _list):
         assert len(_list)
